requirements.txt
torch==1.6.0
GitPython==3.0.5
gitdb2==2.0.6
rlcard
